Move some of the default localization initializers to the SpectatorView script
